backend/requirements.txt
# /send_me_mvp/backend/requirements.txt

# Core FastAPI and ASGI Server
fastapi
uvicorn[standard]
pydantic

# Database: PostgreSQL (for Cloud SQL)
# נשתמש ב-asyncpg לפיתוח async, אך עבור MVP נתחיל עם psycopg2-binary
sqlalchemy # ORM/SQL Toolkit
psycopg2-binary # מנהל התקן ל-PostgreSQL

# LLM Engine
openai

# Google Cloud Services
google-cloud-storage # GCS (לאחסון קורות חיים)
google-cloud-vision # OCR (למודעות משרה)
google-auth-oauthlib # נדרש ל-Gmail API OAuth Flow

# Utilities
python-dotenv # לטעינת משתני סביבה לוקאלית